# src/main/java/com/borzyshka/devicepool/controller/dto/response/BundlePoolReservationResponse.java
package com.borzyshka.devicepool.controller.dto.response;

import lombok.NonNull;

import java.util.List;

public record BundlePoolReservationResponse(@NonNull List<BundleReservationResponse> bundles) {
}