# src/main/java/com/borzyshka/devicepool/controller/dto/response/ServerPoolStatusResponse.java
package com.borzyshka.devicepool.controller.dto.response;

import lombok.NonNull;

import java.util.List;

public record ServerPoolStatusResponse(int size, @NonNull List<ServerPoolObjectStatusResponse> objects) {
}